Color texture experiment
Added a third drawing pass for rendering the light color, the drawing passes are now referenced by enum.
At the moment the color texture overrides the default light texture, will be applied to the correct position (lower half) in the texture later.
The light texture is cleared with white color. This is intended for the lower half. Scissor stuff in comments for now (yes, placeholders like that are bad style)