Dynamic color
The light color is now the color that was set by the user, instead of a default color as in the previous implementation.
The blending of light from multiple colored light sources works correctly with alpha blending.
The light color is rendered to the lower half of the texture now. The shader accesses the brightness/direction information and color information correctly.